modity authenticator (can't compile)
