+++ src/groups/mqb/mqba/mqba_authenticator.h
@@ -57,9 +57,9 @@ class Dispatcher;
 
 namespace mqba {
 
-// =======================
+// ===================
 // class Authenticator
-// =======================
+// ===================
 
 /// Authenticator for a BlazingMQ session with client or broker
 class Authenticator : public mqbnet::Authenticator {
@@ -67,46 +67,8 @@ class Authenticator : public mqbnet::Authenticator {
     // TYPES
 
   private:
-    // PRIVATE TYPES
-    struct ConnectionType {
-        // Enum representing the type of session being negotiated, from that
-        // side of the connection's point of view.
-        enum Enum {
-            e_UNKNOWN,
-            e_CLUSTER_PROXY,   // Reverse connection proxy -> broker
-            e_CLUSTER_MEMBER,  // Cluster node -> cluster node
-            e_CLIENT,          // Either SDK or Proxy -> Proxy or cluster node
-            e_ADMIN
-        };
-    };
-
-    /// Struct used to hold the context associated to a session being
-    /// negotiated
-    struct AuthenticationContext {
-        // PUBLIC DATA
-
-        /// The associated authenticatorContext, passed in by the caller.
-        mqbnet::AuthenticatorContext* d_authenticatorContext_p;
-
-        /// The channel to use for the authentication.
-        bsl::shared_ptr<bmqio::Channel> d_channelSp;
-
-        /// The callback to invoke to notify of the status of the
-        /// authentication.
-        mqbnet::Authenticator::AuthenticationCb d_authenticationCb;
-
-        /// The negotiation message received from the remote peer.
-        bmqp_ctrlmsg::NegotiationMessage d_authenticationMessage;
-
-        /// True if this is a "reversed" connection (on either side of the
-        /// connection).
-        bool d_isReversed;
-
-        /// The type of the session being negotiated.
-        ConnectionType::Enum d_connectionType;
-    };
-
-    typedef bsl::shared_ptr<AuthenticationContext> AuthenticationContextSp;
+    typedef bsl::shared_ptr<mqbnet::AuthenticationContext>
+        AuthenticationContextSp;
 
   private:
     // DATA
@@ -127,33 +89,14 @@ class Authenticator : public mqbnet::Authenticator {
   private:
     // PRIVATE MANIPULATORS
 
-    /// Read callback method invoked when receiving data in the specified
-    /// `blob`, if the specified `status` indicates success.  The specified
-    /// `numNeeded` can be used to indicate if more bytes are needed in
-    /// order to get a full message.  The specified `context` holds the
-    /// negotiation context associated to this read.
-    void readCallback(const bmqio::Status&           status,
-                      int*                           numNeeded,
-                      bdlbb::Blob*                   blob,
-                      const AuthenticationContextSp& context);
-
-    /// Decode the negotiation messages received in the specified `blob` and
-    /// store it, on success, in the corresponding member of the specified
-    /// `context`, returning 0.  Return a non-zero code on error and
-    /// populate the specified `errorDescription` with a description of the
-    /// error.
-    int decodeNegotiationMessage(bsl::ostream& errorDescription,
-                                 const AuthenticationContextSp& context,
-                                 const bdlbb::Blob&             blob);
-
-    /// Invoked when received a `ClientIdentity` negotiation message with
+    /// Invoked when received a `ClientIdentity` authentication message with
     /// the specified `context`.  Creates and return a Session on success,
     /// or return a null pointer and populate the specified
     /// `errorDescription` with a description of the error on failure.
     int onAuthenticationRequest(bsl::ostream& errorDescription,
                                 const AuthenticationContextSp& context);
 
-    /// Invoked when received a `BrokerResponse` negotiation message with
+    /// Invoked when received a `BrokerResponse` authentication message with
     /// the specified `context`.  Creates and return a Session on success,
     /// or return a null pointer and populate the specified
     /// `errorDescription` with a description of the error on failure.
@@ -164,21 +107,17 @@ class Authenticator : public mqbnet::Authenticator {
     /// specified `context` and return 0 on success, or return a non-zero
     /// code on error and populate the specified `errorDescription` with a
     /// description of the error.
-    int
-    sendAuthenticationMessage(bsl::ostream& errorDescription,
-                              const bmqp_ctrlmsg::NegotiationMessage& message,
-                              const AuthenticationContextSp&          context);
+    int sendAuthenticationMessage(
+        bsl::ostream&                              errorDescription,
+        const bmqp_ctrlmsg::AuthenticationMessage& message,
+        const AuthenticationContextSp&             context);
 
-    /// Initiate an outbound negotiation (i.e., send out some negotiation
+    /// Initiate an outbound authentication (i.e., send out some authentication
     /// message and schedule a read of the response) using the specified
     /// `context`.
     /// Senario: reverse connection
     void
     initiateOutboundAuthentication(const AuthenticationContextSp& context);
-
-    /// Schedule a read for the negotiation of the session of the specified
-    /// `context`.
-    void scheduleRead(const AuthenticationContextSp& context);
 
   public:
     // TRAITS
@@ -205,25 +144,19 @@ class Authenticator : public mqbnet::Authenticator {
     // MANIPULATORS
     //   (virtual: mqbnet::Authenticator)
 
-    /// Negotiate the connection on the specified `channel` associated with
-    /// the specified negotiation `context` and invoke the specified
-    /// `negotiationCb` once the negotiation is complete (either success or
-    /// failure).  Note that if no negotiation are needed, the
-    /// `negotiationCb` may be invoked directly from inside the call to
-    /// `negotiate`.
-    void authenticate(mqbnet::AuthenticatorContext*          context,
-                      const bsl::shared_ptr<bmqio::Channel>& channel,
-                      const mqbnet::Authenticator::AuthenticationCb&
-                          authenticationCb) BSLS_KEYWORD_OVERRIDE;
+    /// Send out outbound authentication message or reverse connection request
+    /// with the specified `context`.
+    int authenticationOutboundOrReverse(const AuthenticationContextSp& context)
+        BSLS_KEYWORD_OVERRIDE;
 };
 
 // ============================================================================
 //                             INLINE DEFINITIONS
 // ============================================================================
 
-// -----------------------
+// -------------------
 // class Authenticator
-// -----------------------
+// -------------------
 
 inline Authenticator&
 Authenticator::setClusterCatalog(mqbblp::ClusterCatalog* value)
